Resolve circular import with validating password and renamed _schema to output_schema
