+++ vantage6-server/vantage6/server/resource/collaboration.py
@@ -17,7 +17,7 @@ from vantage6.server.permission import (
     Operation as P,
     PermissionManager
 )
-from vantage6.server.resource.common._schema import (
+from vantage6.server.resource.common.output_schema import (
     CollaborationSchema,
     TaskSchema,
     OrganizationSchema,
