+++ vantage6-server/vantage6/server/resource/common/auth_helper.py
@@ -1,16 +1,13 @@
 import logging
 import datetime as dt
 import pyotp
-import re
 
 from http import HTTPStatus
 from flask import request, render_template
 from flask_mail import Mail
 
-
 from vantage6.common.globals import APPNAME, MAIN_VERSION_NAME
 from vantage6.server.globals import DEFAULT_SUPPORT_EMAIL_ADDRESS
-from vantage6.server import db
 from vantage6.server.model.user import User
 
 module_name = __name__.split('.')[-1]
@@ -19,7 +16,7 @@ log = logging.getLogger(module_name)
 
 def user_login(
     config: dict, username: str, password: str, mail: Mail
-) -> tuple[dict | db.User, HTTPStatus]:
+) -> tuple[dict | User, HTTPStatus]:
     """
     Returns user a message in case of failed login attempt.
 
@@ -43,8 +40,8 @@ def user_login(
     """
     log.info(f"Trying to login '{username}'")
     failed_login_msg = "Failed to login"
-    if db.User.username_exists(username):
-        user = db.User.get_by_username(username)
+    if User.username_exists(username):
+        user = User.get_by_username(username)
         password_policy = config.get("password_policy", {})
         max_failed_attempts = password_policy.get('max_failed_attempts', 5)
         inactivation_time = password_policy.get('inactivation_minutes', 15)
@@ -72,44 +69,8 @@ def user_login(
     return {"msg": failed_login_msg}, HTTPStatus.UNAUTHORIZED
 
 
-def validate_password(pw: str) -> None:
-    """
-    Check if the password meets the password policy requirements
-
-    Parameters
-    ----------
-    pw: str
-        Password to be validated
-
-    Raises
-    ------
-    ValueError
-        If the password does not meet the password policy requirements
-    """
-    if len(pw) < 8:
-        raise ValueError(
-            "Password too short: use at least 8 characters with mixed "
-            "lowercase, uppercase, numbers and special characters"
-        )
-    elif len(pw) > 128:
-        # because long passwords can be used for DoS attacks (long pw
-        # hashing consumes a lot of resources)
-        raise ValueError("Password too long: use at most 128 characters")
-    elif re.search('[0-9]', pw) is None:
-        raise ValueError("Password should contain at least one number")
-    elif re.search('[A-Z]', pw) is None:
-        raise ValueError(
-            "Password should contain at least one uppercase letter")
-    elif re.search('[a-z]', pw) is None:
-        raise ValueError(
-            "Password should contain at least one lowercase letter")
-    elif pw.isalnum():
-        raise ValueError(
-            "Password should contain at least one special character")
-
-
 def notify_user_blocked(
-    user: db.User, max_n_attempts: int, min_rem: int, mail: Mail,
+    user: User, max_n_attempts: int, min_rem: int, mail: Mail,
     config: dict
 ) -> None:
     """
