+++ vantage6-server/vantage6/server/resource/common/output_schema.py
