+++ vantage6-server/vantage6/server/resource/organization.py
@@ -19,7 +19,7 @@ from vantage6.server.resource.common.input_schema import (
 from vantage6.server.resource import (
     with_user_or_node, only_for, with_user, ServicesResources
 )
-from vantage6.server.resource.common._schema import (
+from vantage6.server.resource.common.output_schema import (
     OrganizationSchema,
     CollaborationSchema,
     NodeSchema
