+++ vantage6-server/vantage6/server/resource/task.py
@@ -16,7 +16,7 @@ from vantage6.server.permission import (
     Operation as P
 )
 from vantage6.server.resource import only_for, ServicesResources, with_user
-from vantage6.server.resource.common._schema import (
+from vantage6.server.resource.common.output_schema import (
     TaskSchema,
     TaskIncludedSchema,
 )
